FIX: Improve Reading File Time
The most notable way to decrease time execution was to move out function that calculate user ip from string to digit and pass it as an argument. It was calculated on each line reading which was unnecessary. The final time execution was decrease from around 500ms to around 300ms

FIX# 3